refactoring server and error handling in client

+++ server/src/api/v1/controllers/User_management_controller/user_addressController.ts
@@ -1,20 +1,23 @@
 import expressAsyncHandler from "express-async-handler";
 import { Request, NextFunction, Response } from "express";
 import user_addressModel from "../../models/User_management/user_addressModel";
-
+import { NotFound, MongoDBError } from "../../interfaces/ErrorInstances";
 export const createUserAddress = expressAsyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
     try {
-      console.log(req.body);
       const userAddress = new user_addressModel(req.body);
       await userAddress.save();
-
       res.status(201).json({
         message: "User address created successfully",
         data: userAddress,
       });
-    } catch (error) {
-      next(error);
+    } catch (error: any) {
+      if (error.name === "ValidationError") {
+        const errors = Object.values(error.errors).map(
+          (error: any) => error.message
+        );
+        next(new MongoDBError(400, errors));
+      } else next(error);
     }
   }
 );
@@ -23,6 +26,7 @@ export const updateUserAddress = expressAsyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
     try {
       const { address_id } = req.params;
+      if (!address_id) throw new NotFound(404, "address_id not found");
       const {
         user_id,
         province_code,
@@ -30,6 +34,12 @@ export const updateUserAddress = expressAsyncHandler(
         ward_code,
         address_line_option,
       } = req.body;
+      if (!user_id || !province_code || !district_code || !ward_code)
+        throw new NotFound(
+          404,
+          "Missing some of requirements to create address"
+        );
+
       const userAddress = await user_addressModel.findByIdAndUpdate(
         address_id,
         {
@@ -43,7 +53,7 @@ export const updateUserAddress = expressAsyncHandler(
       );
 
       if (!userAddress) {
-        throw new Error("User address not found");
+        throw new NotFound(404, "User address not found");
       }
 
       res.status(200).json({
@@ -60,10 +70,10 @@ export const deleteUserAddress = expressAsyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
     try {
       const { address_id } = req.params;
+      if (!address_id) throw new NotFound(404, "address_id not found");
       const userAddress = await user_addressModel.findByIdAndDelete(address_id);
-
       if (!userAddress) {
-        throw new Error("User address not found");
+        throw new NotFound(404, "User address not found");
       }
 
       res.status(200).json({
